dashboard/requirements.txt
# DeepEarth Dashboard Requirements
# Python 3.8+ required

# Web Framework
flask>=2.3.0
gunicorn>=20.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For reading Parquet files

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.3
hdbscan>=0.8.0

# Visualization
matplotlib>=3.7.0
pillow>=10.0.0

# 3D Data Processing
open3d>=0.17.0

# Optional but recommended
tqdm>=4.65.0  # Progress bars for data conversion
huggingface-hub>=0.19.0  # For downloading complete datasets from HuggingFace